feat: 프론트 변경사항 반영

+++ challenger/src/main/java/academy/challenger/challenge/ChallengeResponse.java
@@ -5,7 +5,7 @@ public record ChallengeResponse (
         long userId,
         String title,
         String description,
-        Integer progress,
+        Double progress,
         Integer duration
 ) {
 }
